Created Shields
Added shields component which will take damage before the health does. If the shield takes more damage than the shield value, the remaining damage is applied to the player health.

Created a shield UI bar for both the player and the enemy, they work the same way but look different on the ui. When the shield value changes an event is fired that alerts the UI to update itself to the new value.